+++ tests/test_parser_errors.py
@@ -2,17 +2,18 @@ import io
 
 import pytest
 
-from . import TEST_DIR
+from cp2k_input_tools import DEFAULT_CP2K_INPUT_XML
 from cp2k_input_tools.parser import CP2KInputParser
-from cp2k_input_tools.cli import DEFAULT_CP2K_INPUT_XML
 from cp2k_input_tools.parser_errors import (
-    InvalidParameterError,
-    PreprocessorError,
     InvalidNameError,
+    InvalidParameterError,
     InvalidSectionError,
+    PreprocessorError,
     SectionMismatchError,
 )
 from cp2k_input_tools.tokenizer import UnterminatedStringError
+
+from . import TEST_DIR
 
 
 def test_error_invalid_number_of_parameters():
@@ -23,7 +24,7 @@ def test_error_invalid_number_of_parameters():
             cp2k_parser.parse(fhandle)
 
     assert "invalid values for keyword: A" in excinfo.value.args[0]
-    assert excinfo.value.args[1]["linenr"] == 41
+    assert excinfo.value.args[1].linenr == 41
     assert isinstance(excinfo.value.__cause__, InvalidParameterError)
     assert "keyword expects exactly 3 values, 2 were given" in excinfo.value.__cause__.args[0]
 
@@ -35,7 +36,7 @@ def test_unterminated_string():
         with pytest.raises(UnterminatedStringError) as excinfo:
             cp2k_parser.parse(fhandle)
 
-    assert excinfo.value.args[1]["linenr"] == 14
+    assert excinfo.value.args[1].linenr == 14
 
 
 def test_undefined_preprocessor_var():
@@ -46,7 +47,7 @@ def test_undefined_preprocessor_var():
             cp2k_parser.parse(fhandle)
 
     assert "undefined variable 'HP'" in excinfo.value.args[0]
-    assert excinfo.value.args[1]["linenr"] == 30
+    assert excinfo.value.args[1].linenr == 30
 
 
 def test_multiple_defined_non_repeating_section():