+++ tests/test_parser_preprocessor.py
@@ -2,10 +2,11 @@ import io
 
 import pytest
 
-from . import TEST_DIR
+from cp2k_input_tools import DEFAULT_CP2K_INPUT_XML
 from cp2k_input_tools.parser import CP2KInputParser
-from cp2k_input_tools.cli import DEFAULT_CP2K_INPUT_XML
 from cp2k_input_tools.parser_errors import PreprocessorError
+
+from . import TEST_DIR
 
 
 def test_var_substition():